backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
groq
python-dotenv
python-multipart
